[vulkan_rhi, renderer]: Temporarily fix synchronization issues
Currently the frame graph does a poor job of inserting barriers, and should be rewritten to a DAG(directed acyclic graph) anyway.
It would also be a good idea to split `AccessFlags` into `BufferAccessFlags`, `TextureAccessFlags` and `MemoryAccessFlags`. This would allow to improve the granularity of the barriers and simplify things a bit.